Support arbitrary third party prefixes (on 2bc93b1c)
`goimports` currently sorts imports into two main groups - one for
imports from the stdlib and another for all other imports.  This is
largely disguished by whether or not the first component of the import
has a "." (period) in it or not.

Occasionally, there'll be import paths where the first component of the
import does not contain a "." and is not part of the stdlib.  Provide an
option to categorize these import paths as not part of the stdlib, but
are infact a 3rd party import.